Removing deadlock in Consul manager
The code in `continuouslyAcquireConsulLeadership()` had a logic bomb that
could cause any new Consul state changes to not propagate to the reconciler.

`continuouslyAcquireConsulLeadership` used to `select` on two channels in an infinite loop

* `timer.C`: Periodically would call `acquireConsulLeadership` every 250ms
* `m.masterCh`: The intend for the channel was to quickly react to leadership changes and try to acquire it, rather than letting the `timer` trigger the work. This was put in place to decrease recovery time for new leadership election.

`acquireConsulLeadership` setups a Consul Lock and tries to acquire the mutex (meaning the Resec process is now leader).

It was assumed at the time that `acquireConsulLeadership` would `return` in case it could not acquire the mutex within a reasonable time, and in those cases, the `m.masterCh` would be drained in `continuouslyAcquireConsulLeadership` until the next `timer.C` triggered.

But since this is not the case the `m.masterCh` channel would continue to get new writes from `watchConsulMasterService`, until the buffered channel is full, and end up blocking the function - since a full channel will block writes.

Responsible for following the IP+Port of the currently elected Resec leader (decided via Consul Lock mutex in `acquireConsulLeadership`).

When a leadership change happened, it would update the reconciler state and then write an update to the `m.masterCh` to manually trigger a faster leadership reelection.

However, when the writes to `m.masterCh` blocks, it also mean that the reconciler no longer will get any Consul leadership change updates (since they are emitted in the same loop as `m.masterCh` writes that now blocks), so it will be unaware of a new leader having been selected in the Resec culster, causing the process to never recover on its own.

Signed-off-by: Christian Winther <jippignu@gmail.com>